requirements.txt
discord.py
python-dotenv
pytz
asyncio
